+++ backend/src/server/cors.ts
@@ -1,0 +1,8 @@
+import cors from 'cors'
+
+const corsOptions = {
+  origin: '*',
+  exposedHeaders: ['token'],
+}
+
+export default cors(corsOptions)
