+++ backend/src/server/plugins.ts
@@ -1,0 +1,17 @@
+const plugins = [
+  {
+    requestDidStart() {
+      return {
+        willSendResponse(requestContext: any) {
+          const { setHeaders = [] } = requestContext.context
+          setHeaders.forEach(({ key, value }: { [key: string]: string }) => {
+            requestContext.response.http.headers.append(key, value)
+          })
+          return requestContext
+        },
+      }
+    },
+  },
+]
+
+export default plugins
